Add initial version
A basic Django app for managing shared cash boxes is provided.
Cash can flow between an user and a cash box and is described as a
transaction. If a transaction is specialized as an invoice, the amount
adds to the general balance of the cash box.
Therefore double-entry bookkeeping and a public log for expenses
are both realized at the same time.
A normal user should only be able to add invoices.
The treasurer can additionally add transactions (describing cash flow).